backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
openpyxl==3.1.2

django-storages[boto3]==1.14.4
boto3==1.35.14
dj-database-url==3.0.1
